algorithms

+++ algorithms/algorithms-with-iterators/task.h
@@ -147,7 +147,7 @@ public:
     FibonacciRange(size_t amount) {
         amount_= amount;
         int64_t prev1 = 1, prev2 = 2;
-        for (int i = 3; i <= amount + 2; ++i) {
+        for (int i = 3; i <= (int) amount + 2; ++i) {
             int64_t help = prev2;
             prev2 = prev1 + prev2;
             prev1 = help;
